chore(keylog): separate TLS keylog responsibilities
